backend/requirements.txt
django>=4.2
djangorestframework
django-cors-headers
psycopg2-binary
langchain
langchain-core
langchain-groq
langgraph
langchain-mcp-adapters
python-dotenv
channels
PyJWT
cryptography
requests